// VIS_Application/app/Model/HumanResource/ProfileAttendance/MyProfiles.ts
export interface IMyProfiles {
    Mode: string;
    UserId: number;
    Employee_Name: string;
    LineManagerName: string;
    FirstName: string;
    MiddleName: string;
    LastName: string;
    Gender: string;
    PhotographFileName: string;
    CompanyName: string;
    Email: string;
    Designation: string;
    DepartmentName: string;
    FatherName: string;
    MotherName: string;
    PermenantAddres: string;
    CommunicationAddress: string;
    LandlineNumber: string;
    MobileNumber: string;
    BirthDate: string;
    BloodGroup: string;
    MaritalStatus: string;

    ClassDegree: string;
    SchoolInstitute: string;
    Medium: string;
    PassingYear: number;
    BoardUniversity: string;
    Percentage: string;


    
    //Employee_Id: number;



}